+++ src/core/request.ml
@@ -183,10 +183,10 @@ let toplevel_metadata t =
     | (h :: _) :: _ -> h.metadata
 
 let iter_metadata t f =
+  f t.root_metadata;
   List.iter
     (function [] -> assert false | h :: _ -> f h.metadata)
-    t.indicators;
-  f t.root_metadata
+    t.indicators
 
 let set_metadata t k v = Hashtbl.replace (toplevel_metadata t) k v
 let set_root_metadata t k v = Hashtbl.replace t.root_metadata k v
@@ -197,14 +197,8 @@ let get_metadata t k =
   try
     iter_metadata t (fun h ->
         try raise (Found (Hashtbl.find h k)) with Not_found -> ());
-    (try raise (Found (Hashtbl.find t.root_metadata k)) with Not_found -> ());
     None
   with Found s -> Some s
-
-let get_root_metadata t k =
-  try raise (Found (Hashtbl.find t.root_metadata k)) with
-    | Not_found -> None
-    | Found x -> Some x
 
 let get_all_metadata t =
   let h = Hashtbl.create 20 in
@@ -470,10 +464,6 @@ let get_metadata t k =
 let get_all_metadata t =
   update_metadata t;
   get_all_metadata t
-
-let get_root_metadata t =
-  update_metadata t;
-  get_root_metadata t
 
 (** Global management *)