+++ CHANGES.md
@@ -8,6 +8,7 @@ New:
 
 Fixed:
 
+- Prevent request metadata from overriding root metadata (#3813)
 - Fixed `source.drop` and `source.dump` clock initialization.
 - Fixed bogus report of non-monotonous PTS content when using raw ffmpeg content.
 - Fixed streaming errors when disconnecting `input.harbor`.